Had to revert namespace stuff to get PackedVector building on Win32 (looks like the death knell in making this portable; time to investigate isolating the bits I need)
Lua range iterator now uses int rather than size_t owing to typical use

Relaxed parallel_for type

+++ utils/LuaEx.cpp
@@ -393,7 +393,7 @@ namespace LuaXS {
 		return *this;
 	}
 
-	Range::Iter::operator size_t (void)
+	Range::Iter::operator int (void)
 	{
 		return mIndex + 1;
 	}
@@ -411,7 +411,7 @@ namespace LuaXS {
 		{
 			Iter ei{*this};
 
-			ei.mIndex = 0U;
+			ei.mIndex = 0;
 
 			return ei;
 		}
@@ -423,9 +423,9 @@ namespace LuaXS {
 	{
 		Iter ei{*this};
 
-		if (mIsTable) ei.mIndex = lua_objlen(mL, mArg);
+		if (mIsTable) ei.mIndex = int(lua_objlen(mL, mArg));
 
-		else ei.mIndex = mNonTableOK ? 1U : 0U;
+		else ei.mIndex = mNonTableOK ? 1 : 0;
 
 		return ei;
 	}
